Refactor & rename lastfm_service to use Melon
